Backend: Fossil Excavator API (#1431)
