+++ src/main/java/at/hannibal2/skyhanni/features/mining/fossilexcavator/solver/FossilType.kt
@@ -1,4 +1,4 @@
-package at.hannibal2.skyhanni.features.mining.fossilexcavator
+package at.hannibal2.skyhanni.features.mining.fossilexcavator.solver
 
 enum class FossilType(
     val displayName: String,
